Removed elastic test-runner script.
The tests fail erratically if you run them on a freshly spun up elasticsearch
instance, but if you run them on the instance you've been using for dev they
pass.

The app also works.

We'll eventually need to find a way to run this as part of CI, but not just
now.